cleaned up some code, made it less congested
